+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## [Unreleased]
 
+* *Breaking*: Use a generic type to support pin checking on SDRAMs with 11 and
+  13 address lines. `PinsSdram` now has two generic types.
+
 ## [v0.1.2] 2020-08-05
 
 * Don't require type to be `Sync` in order to implement FmcPeripheral